@@ -1,6 +1,6 @@
 [tool.black]
-line-length = 70
-target-version = ['py39']
+line-length = 80
+target-version = ['py310']
 include = '\.pyi?$'
 extend-exclude = '''
 # A regex preceded with ^/ will apply only to files and directories
@@ -12,9 +12,9 @@ extend-exclude = '''
 profile = "black"
 atomic = true
 group_by_package = true
-py_version = 39
+py_version = 310
 skip = [".gitignore", ".dockerignore"]
-line_length = 70
+line_length = 80
 sections = [
   'FUTURE',
   'STDLIB',
@@ -30,17 +30,26 @@ known_avalanche = ['avalanche']
 known_doit = ['doit']
 
 [tool.mypy]
+python_version = "3.10"
 files = ["src", "workflow", "*.py"]
+# exclude = 'wandb|tests|OLD_FILES|src/utils/summary.py|src/dataset/google/.*\.py'
 exclude = [
   '^wandb',
   '^tests',
   '^OLD_FILES',
   '^src/utils/summary.py',
+  '^workflow',
+  '^src/dataset/google/.*\.py', # NOTE: @william remove this after migration
 ]
 plugins = [
-  # "pydantic.mypy"
+  "pydantic.mypy"
 ]
 ignore_missing_imports = true
+
+[[tool.mypy.overrides]]
+module = "src.google.dataset.*"
+ignore_missing_imports = true
+follow_imports = "skip"
 # follow_imports = "silent"
 # warn_redundant_casts = true
 # warn_unused_ignores = true
